Move func vm value conversion to generators folder
