+++ kybra/compiler/kybra_generate/src/generators/mod.rs
@@ -1,6 +1,7 @@
 pub mod async_result_handler;
 pub mod body;
 pub mod canister_methods;
+pub mod func;
 pub mod header;
 pub mod ic_object;
 pub mod kybra_serde;
